only execute if exists

+++ machines/rbair/uninstall.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
 # stop kimnaty and remove services etc.
-pushd "/home/pi/kimnaty" || exit 0
-  su -c "/home/pi/kimnaty/kimnaty --uninstall" pi
-  # remove rclone's local mountpoint
-  sudo rm -rv /srv/rmt
-popd || exit 0
+if [ -d /home/pi/kimnaty ]; then
+  pushd "/home/pi/kimnaty" || exit 0
+    su -c "/home/pi/kimnaty/kimnaty --uninstall" pi
+    # remove rclone's local mountpoint
+    sudo rm -rv /srv/rmt
+  popd || exit 0
+fi
