+++ machines/rbelec/uninstall.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
 # stop kimnaty and remove services etc.
-pushd "/home/pi/lektrix" || exit 0
-  su -c "/home/pi/lektrix/lektrix --uninstall" pi
-   # remove rclone's local mountpoint
-  sudo rm -rv /srv/rmt
+if [ -d /home/pi/lektrix ]; then
+  pushd "/home/pi/lektrix" || exit 0
+    su -c "/home/pi/lektrix/lektrix --uninstall" pi
+     # remove rclone's local mountpoint
+    sudo rm -rv /srv/rmt
 popd || exit 0
+fi
